refactor: move type StoreStatus and Node to store package
